always use the specified dispatcher in builder functions
